src/video-intelligence/requirements.txt
functions-framework==3.*
google-cloud-videointelligence==2.12.0
google-cloud-aiplatform
google-cloud-storage
pandas
MoviePy
pandas-gbq
tenacity